# gestimowebfront-end/src/tools/swagger/src/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Swagger API pour GESTIMOWEB REST API",
    "version" : "0.0.1-SNAPSHOT",
    "title" : "Swagger GESTIMOWEB REST API",
    "contact" : {
      "name" : "bzadata",
      "email" : "michel.bossoh@outlook.fr"
    }
  },
  "basePath" : "http://localhost:8282/",
  "paths" : {
...
    "gestimoweb/api/v1/prestation/delete/{id}" : {
      "delete" : {
        "operationId" : "deleteServiceAdditionnelPrestation",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "gestimoweb/api/v1/prestation/findById/{id}" : {
      "get" : {
        "operationId" : "findServiceAdditionnelByIDPrestation",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/PrestationSaveOrUpdateDto"
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/prestation/saveorupdate" : {
      "post" : {
        "operationId" : "saveorupdatePrestation",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/PrestationSaveOrUpdateDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/PrestationSaveOrUpdateDto"
            }
          }
        }
      }
    },
...
    "gestimoweb/api/v1/prixparcategorie/listDesPrixParCategori/{idCategorie}" : {
      "get" : {
        "operationId" : "listDesPrixParCategori",
        "parameters" : [ {
          "name" : "idCategorie",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/PrixParCategorieChambreDto"
              }
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/prixparcategorie/saveOrUpDatePrixParCategorie" : {
      "post" : {
        "operationId" : "saveOrUpDatePrixParCategorie",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/PrixParCategorieChambreDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/PrixParCategorieChambreDto"
            }
          }
        }
      }
    },
...
    "gestimoweb/api/v1/reservation/allreservationparagence/{idAgence}" : {
      "get" : {
        "operationId" : "allreservationparagence",
        "parameters" : [ {
          "name" : "idAgence",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ReservationAfficheDto"
              }
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/reservation/delete/{id}" : {
      "delete" : {
        "operationId" : "deleteReservation",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "gestimoweb/api/v1/reservation/findAllEncaissementReservationByIdBien/{idReser}" : {
      "get" : {
        "operationId" : "findAllEncaissementReservationByIdBien",
        "parameters" : [ {
          "name" : "idReser",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/EncaissementReservationDto"
              }
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/reservation/findPeriodeReservationByIdBien/{idBien}" : {
      "get" : {
        "operationId" : "findPeriodeReservationByIdBien",
        "parameters" : [ {
          "name" : "idBien",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ReservationAfficheDto"
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/reservation/findReservationById/{id}" : {
      "get" : {
        "operationId" : "findCategorieChambreByIDReservation",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ReservationAfficheDto"
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/reservation/listeDesReservationOuvertParAgence/{idAgence}" : {
      "get" : {
        "operationId" : "listeDesReservationOuvertParAgence",
        "parameters" : [ {
          "name" : "idAgence",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ReservationAfficheDto"
              }
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/reservation/saveencaissementreservation" : {
      "post" : {
        "operationId" : "saveencaissementreservation",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/EncaissementReservationRequestDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/EncaissementReservationDto"
              }
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/reservation/saveorupdate" : {
      "post" : {
        "operationId" : "saveorupdateRes",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/ReservationSaveOrUpdateDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ReservationSaveOrUpdateDto"
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/reservation/saveorupdategood" : {
      "post" : {
        "operationId" : "saveorupdategood",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/ReservationRequestDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ReservationAfficheDto"
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/reservation/saveorupdatereservation" : {
      "post" : {
        "operationId" : "saveorupdatereservation",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/ReservationRequestDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "EncaissementReservationDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idReservation" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idCreateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "creationDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "modePaiement" : {
          "type" : "string"
        },
        "operationType" : {
          "type" : "string"
        },
        "dateEncaissement" : {
          "type" : "string",
          "format" : "date"
        },
        "idLastEncaissement" : {
          "type" : "integer",
          "format" : "int64"
        },
        "montantEncaissement" : {
          "type" : "number",
          "format" : "double"
        },
        "soldeEncaissement" : {
          "type" : "number",
          "format" : "double"
        },
        "encienSoldReservation" : {
          "type" : "number",
          "format" : "double"
        },
        "nvoSoldeReservation" : {
          "type" : "number",
          "format" : "double"
        },
        "intituleDepense" : {
          "type" : "string"
        },
        "entiteOperation" : {
          "type" : "string",
          "enum" : [ "MAGISER", "SEVEINVEST" ]
        },
        "reservationAfficheDto" : {
          "$ref" : "#/definitions/ReservationAfficheDto"
        },
        "typePaiement" : {
          "type" : "string"
        }
      }
    },
    "EncaissementReservationRequestDto" : {
      "type" : "object",
      "properties" : {
        "idReservation" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idCreateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "modePaiement" : {
          "type" : "string"
        },
        "dateEncaissement" : {
          "type" : "string",
          "format" : "date"
        },
        "montantEncaissement" : {
          "type" : "number",
          "format" : "double"
        },
        "encienSoldReservation" : {
          "type" : "number",
          "format" : "double"
        },
        "nvoSoldeReservation" : {
          "type" : "number",
          "format" : "double"
        },
        "idAppartement" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
...
    "PrestationSaveOrUpdateDto" : {
      "type" : "object",
      "required" : [ "name" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idCreateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        },
        "amount" : {
          "type" : "number",
          "format" : "double"
        }
      }
    },
    "PrixParCategorieChambreDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idCreateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "nombreDeJour" : {
          "type" : "string"
        },
        "prix" : {
          "type" : "number",
          "format" : "double"
        },
        "intervalPrix" : {
          "type" : "integer",
          "format" : "int32"
        },
        "description" : {
          "type" : "string"
        },
        "nbrDiffJour" : {
          "type" : "integer",
          "format" : "int32"
        },
        "idCategorieChambre" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
...
    "ReservationAfficheDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idCreateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "creationDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "lastModifiedDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idUtilisateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "email" : {
          "type" : "string"
        },
        "mobile" : {
          "type" : "string"
        },
        "username" : {
          "type" : "string"
        },
        "idAppartementdDto" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idLastEncaissement" : {
          "type" : "integer",
          "format" : "int64"
        },
        "dateDebut" : {
          "type" : "string",
          "format" : "date"
        },
        "dateFin" : {
          "type" : "string",
          "format" : "date"
        },
        "utilisateurOperation" : {
          "type" : "string"
        },
        "bienImmobilierOperation" : {
          "type" : "string"
        },
        "designationBail" : {
          "type" : "string"
        },
        "abrvCodeBail" : {
          "type" : "string"
        },
        "enCoursBail" : {
          "type" : "boolean"
        },
        "archiveBail" : {
          "type" : "boolean"
        },
        "montantCautionBail" : {
          "type" : "number",
          "format" : "double"
        },
        "nbreMoisCautionBail" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nouveauMontantLoyer" : {
          "type" : "number",
          "format" : "double"
        },
        "idBienImmobilier" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idLocataire" : {
          "type" : "integer",
          "format" : "int64"
        },
        "codeAbrvBienImmobilier" : {
          "type" : "string"
        },
        "pourcentageReduction" : {
          "type" : "number",
          "format" : "double"
        },
        "montantReduction" : {
          "type" : "number",
          "format" : "double"
        },
        "soldReservation" : {
          "type" : "number",
          "format" : "double"
        },
        "montantPaye" : {
          "type" : "number",
          "format" : "double"
        },
        "nmbreAdulte" : {
          "type" : "integer",
          "format" : "int32"
        },
        "montantReservation" : {
          "type" : "number",
          "format" : "double"
        },
        "nmbrEnfant" : {
          "type" : "integer",
          "format" : "int32"
        },
        "descriptionCategori" : {
          "type" : "string"
        },
        "nameCategori" : {
          "type" : "string"
        },
        "priceCategori" : {
          "type" : "number",
          "format" : "double"
        },
        "nbrDiffJourCategori" : {
          "type" : "integer",
          "format" : "int32"
        },
        "pourcentReducCategori" : {
          "type" : "number",
          "format" : "double"
        }
      }
    },
    "ReservationRequestDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idCreateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAppartementdDto" : {
          "type" : "integer",
          "format" : "int64"
        },
        "dateDebut" : {
          "type" : "string"
        },
        "dateFin" : {
          "type" : "string"
        },
        "idClient" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idBien" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idUtilisateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "nom" : {
          "type" : "string"
        },
        "prenom" : {
          "type" : "string"
        },
        "username" : {
          "type" : "string"
        },
        "pourcentageReduction" : {
          "type" : "number",
          "format" : "double"
        },
        "montantReduction" : {
          "type" : "number",
          "format" : "double"
        },
        "soldReservation" : {
          "type" : "number",
          "format" : "double"
        },
        "montantPaye" : {
          "type" : "number",
          "format" : "double"
        },
        "montantReservation" : {
          "type" : "number",
          "format" : "double"
        },
        "montantDeReservation" : {
          "type" : "number",
          "format" : "double"
        },
        "nmbreAdulte" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nmbrEnfant" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "ReservationSaveOrUpdateDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idCreateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "utilisateurRequestDto" : {
          "$ref" : "#/definitions/UtilisateurRequestDto"
        },
        "idAppartementdDto" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idUtilisateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "dateDebut" : {
          "type" : "string",
          "format" : "date"
        },
        "dateFin" : {
          "type" : "string",
          "format" : "date"
        },
        "bienImmobilierOperation" : {
          "type" : "string"
        },
        "designationBail" : {
          "type" : "string"
        },
        "abrvCodeBail" : {
          "type" : "string"
        },
        "enCoursBail" : {
          "type" : "boolean"
        },
        "archiveBail" : {
          "type" : "boolean"
        },
        "montantCautionBail" : {
          "type" : "number",
          "format" : "double"
        },
        "nbreMoisCautionBail" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nouveauMontantLoyer" : {
          "type" : "number",
          "format" : "double"
        },
        "idBienImmobilier" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idLocataire" : {
          "type" : "integer",
          "format" : "int64"
        },
        "codeAbrvBienImmobilier" : {
          "type" : "string"
        },
        "pourcentageReduction" : {
          "type" : "number",
          "format" : "double"
        },
        "montantReduction" : {
          "type" : "number",
          "format" : "double"
        },
        "soldReservation" : {
          "type" : "number",
          "format" : "double"
        },
        "montantPaye" : {
          "type" : "number",
          "format" : "double"
        },
        "nmbreAdulte" : {
          "type" : "integer",
          "format" : "int32"
        },
        "montantReservation" : {
          "type" : "number",
          "format" : "double"
        },
        "nmbrEnfant" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
...
    "RoleRequestDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idCreateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "roleName" : {
          "type" : "string"
        },
        "descriptionRole" : {
          "type" : "string"
        }
      }
    },
...
    "UtilisateurRequestDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idCreateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "utilisateurIdApp" : {
          "type" : "string"
        },
        "nom" : {
          "type" : "string"
        },
        "prenom" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "mobile" : {
          "type" : "string"
        },
        "dateDeNaissance" : {
          "type" : "string",
          "format" : "date"
        },
        "lieuNaissance" : {
          "type" : "string"
        },
        "typePieceIdentite" : {
          "type" : "string"
        },
        "numeroPieceIdentite" : {
          "type" : "string"
        },
        "dateDebutPiece" : {
          "type" : "string",
          "format" : "date"
        },
        "dateFinPiece" : {
          "type" : "string",
          "format" : "date"
        },
        "nationalite" : {
          "type" : "string"
        },
        "genre" : {
          "type" : "string"
        },
        "username" : {
          "type" : "string"
        },
        "password" : {
          "type" : "string"
        },
        "profileImageUrl" : {
          "type" : "string"
        },
        "lastLoginDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "lastLoginDateDisplay" : {
          "type" : "string",
          "format" : "date-time"
        },
        "joinDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "roleUsed" : {
          "type" : "string"
        },
        "authorities" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "agenceDto" : {
          "type" : "integer",
          "format" : "int64"
        },
        "roleRequestDto" : {
          "$ref" : "#/definitions/RoleRequestDto"
        },
        "userCreate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idEtablissement" : {
          "type" : "integer",
          "format" : "int64"
        },
        "active" : {
          "type" : "boolean"
        },
        "activated" : {
          "type" : "boolean"
        },
        "nonLocked" : {
          "type" : "boolean"
        }
      }
    },
...
  }
}